organize new registry auth code
